sales appp

+++ app1/models.py
@@ -396,7 +396,10 @@ class ItemOrders(models.Model):
 
     payment_type = models.CharField(max_length=50)
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    quantity = models.PositiveIntegerField()
+    quantity = models.DecimalField(
+        max_digits=10,
+        decimal_places=2
+    )
     amount = models.DecimalField(max_digits=10, decimal_places=2)
 
     client_id = models.CharField(max_length=100)
