+++ sales/models.py
@@ -20,7 +20,10 @@ class Sales(models.Model):
 
     payment_type = models.CharField(max_length=50)
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    quantity = models.PositiveIntegerField()
+    quantity = models.DecimalField(
+        max_digits=10,
+        decimal_places=2
+    )
     amount = models.DecimalField(max_digits=10, decimal_places=2)
 
     client_id = models.CharField(max_length=100)
