+++ sales_return/models.py
@@ -14,7 +14,10 @@ class SalesReturn(models.Model):
     barcode = models.CharField(max_length=100)
 
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    quantity = models.PositiveIntegerField()
+    quantity = models.DecimalField(
+        max_digits=10,
+        decimal_places=2
+    )
     amount = models.DecimalField(max_digits=10, decimal_places=2)
 
     product_remark = models.TextField(blank=True, null=True)  # ✅ NEW
